Added dyanamic dispatch (select the correct function to call at runtime based on CPUID info of available SIMD extensions) for THVector functions for x86
