+++ lib/TH/THTensor.c
@@ -1,6 +1,10 @@
 #include "THAtomic.h"
 #include "THTensor.h"
-#include "THVector.h"
+
+#include "THVectorImpls.h"
+#include "generic/THVectorDispatch.c"
+#include "THGenerateAllTypes.h"
+
 #include "THBlas.h"
 #include "THLapack.h"
 #include "THRandom.h"
